Moved class file to lib folder and updated doc
